requirements.txt
django==2.2.26
djangorestframework==3.13.1
psycopg2==2.8.6
python-dotenv==0.19.2
Pillow==9.0.0
